Bugfix/issue98 (#173)
* Support java.time.Instant and BigInteger

* Correct issue #98

- Correct typo
- Make DatabaseProcessor public

* Enhance documentation of transient property calculator

+++ jpa/odata-jpa-test/src/main/resources/db/migration/V1_0__olingo.sql
@@ -718,7 +718,7 @@ CREATE TABLE "NestedComplex" (
 	"Number" INTEGER NOT NULL,
 	"Figure1" INTEGER,
 	"Figure2" INTEGER,
-	"Figure3" INTEGER,
+	"Figure3" BIGINT,
    PRIMARY KEY ("ID", "Number"));		
    
 insert into "NestedComplex" values('501',1, 1, 1, 1);
